Add access methods to XmlValue
Methods to access children and attribute are now added to XmlValue alongside some Specs for those methods.
Tech debt:
Attributes can have equal names if they are of different namespaces. At the moment the operators for attributes \@ \@? do not work with more than one attribute found with a certain name.